refactor: no circular module graph
