+++ src/client/properties/with.ts
@@ -1,10 +1,11 @@
-import { type ConfigInit, updateContext } from '../Configuration/ConfigInit.js'
+import { Context } from '../../types/context.js'
+import { type ConfigInit } from '../Configuration/ConfigInit.js'
 import { createProperties } from '../helpers.js'
 
 export const withProperties = createProperties((builder, state) => {
   return {
     with: (configInit: ConfigInit) => {
-      return builder(updateContext(state, configInit))
+      return builder(Context.updateContextConfigInit(state, configInit))
     },
   } as any
 })
