+++ src/lib/anyware/PipelineDef/builder.ts
@@ -2,9 +2,9 @@ import type { ConfigManager } from '../../config-manager/__.js'
 import { type Tuple } from '../../prelude.js'
 import type { Extension } from '../Extension/__.js'
 import { Overload } from '../Overload/__.js'
+import { Pipeline } from '../Pipeline/Pipeline.js'
 import type { StepDefinition } from '../StepDefinition.js'
 import type { StepRunner } from '../StepRunner.js'
-import { Pipeline } from './_.js'
 import type { PipelineDefinition } from './__.js'
 import { type Options, resolveOptions } from './Config.js'
