+++ src/lib/grafaid/schema/schema.ts
@@ -14,7 +14,6 @@ import {
   isEnumType,
   isInputObjectType,
   isInterfaceType,
-  isNonNullType,
   isObjectType,
   isUnionType,
 } from 'graphql'
@@ -57,8 +56,14 @@ export {
 } from 'graphql'
 
 export * as Args from './args.js'
+
 export * as CustomScalars from './customScalars.js'
+
 export * from './KindMap/__.js'
+
+export * from './typeGuards.js'
+
+export * from './scalars.js'
 
 export type DeprecatableNodes = GraphQLEnumValue | InputOrOutputField
 
@@ -94,49 +99,7 @@ export type DescribableTypes =
   | GraphQLNamedType
   | InputOrOutputField
 
-type InputFieldLikeTypes = GraphQLArgument | GraphQLInputField
-
-export const isInputFieldLike = (value: object): value is InputFieldLikeTypes => {
-  return `defaultValue` in value
-}
-
-export const isField = (value: object): value is GraphQLField<any, any> | InputFieldLikeTypes => {
-  return isOutputField(value) || isInputFieldLike(value)
-}
-
-export const isOutputField = (value: object): value is GraphQLField<any, any> => {
-  return `args` in value
-}
-
-export const isScalarTypeCustom = (node: GraphQLScalarType): boolean => {
-  return !(node.name in standardScalarTypeNames)
-}
-
-export const standardScalarTypeNames = {
-  String: `String`,
-  ID: `ID`,
-  Int: `Int`,
-  Float: `Float`,
-  Boolean: `Boolean`,
-}
-
-export interface StandardScalarRuntimeTypeMap {
-  String: string
-  ID: string
-  Int: number
-  Float: number
-  Boolean: boolean
-}
-
-export type StandardScalarRuntimeTypes = StandardScalarRuntimeTypeMap[keyof StandardScalarRuntimeTypeMap]
-
-export const isScalarTypeAndCustom = (node: unknown): node is GraphQLScalarType => {
-  return isScalarType(node) && isScalarTypeCustom(node)
-}
-
-export const isAllInputObjectFieldsNullable = (node: GraphQLInputObjectType) => {
-  return Object.values(node.getFields()).some(_ => !isNonNullType(_.type))
-}
+export type InputFieldLikeTypes = GraphQLArgument | GraphQLInputField
 
 export const NameToClassNamedType = {
   GraphQLScalarType: GraphQLScalarType,
